test: Set CT map sizes for Nightly update test
Currently, the Nightly CI run performs the update test from v1.3 to
the latest master. The default CT maps size in v1.3 is 524288, while
in the master - 1000000. The size mismatch breaks the svc migration
test, as due to the mismatch the CT maps get removed and thus, existing
connections get broken.

Signed-off-by: Martynas Pumputis <m@lambda.lt>